refactor: Move memversClient to module directory
